♻️ App.tsxに集中していたロジックの責務分離 UI修正
